Move root level component related classes into the components package
COSDK-500
