+++ core/src/main/java/com/adyen/checkout/core/components/CheckoutCallback.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 16/5/2025.
  */
 
-package com.adyen.checkout.core
+package com.adyen.checkout.core.components
 
 import com.adyen.checkout.core.components.paymentmethod.PaymentComponentState
