+++ core/src/main/java/com/adyen/checkout/core/components/CheckoutConfiguration.kt
@@ -6,13 +6,14 @@
  * Created by ozgur on 2/5/2025.
  */
 
-package com.adyen.checkout.core
+package com.adyen.checkout.core.components
 
 import android.annotation.SuppressLint
 import android.os.Parcel
 import android.os.Parcelable
 import android.os.Parcelable.CONTENTS_FILE_DESCRIPTOR
 import androidx.annotation.RestrictTo
+import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.common.internal.helper.CheckoutConfigurationMarker
 import com.adyen.checkout.core.common.internal.helper.LocaleUtil
