+++ core/src/main/java/com/adyen/checkout/core/internal/PaymentMethodFactory.kt
@@ -8,7 +8,7 @@
 
 package com.adyen.checkout.core.internal
 
-import com.adyen.checkout.core.CheckoutConfiguration
+import com.adyen.checkout.core.components.CheckoutConfiguration
 import com.adyen.checkout.core.internal.ui.PaymentDelegate
 import com.adyen.checkout.core.internal.ui.model.SessionParams
 import kotlinx.coroutines.CoroutineScope
