+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/MBWayConfiguration.kt
@@ -9,10 +9,10 @@
 package com.adyen.checkout.core.mbway.internal.ui
 
 import android.content.Context
-import com.adyen.checkout.core.CheckoutConfiguration
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.common.internal.helper.CheckoutConfigurationMarker
+import com.adyen.checkout.core.components.CheckoutConfiguration
 import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.internal.ActionHandlingPaymentMethodConfigurationBuilder
 import com.adyen.checkout.core.internal.ButtonConfiguration
