+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSession.kt
@@ -9,8 +9,8 @@
 package com.adyen.checkout.core.sessions
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.CheckoutConfiguration
 import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.components.CheckoutConfiguration
 import com.adyen.checkout.core.data.model.PaymentMethod
 import com.adyen.checkout.core.sessions.internal.data.model.SessionSetupResponse
